Move go/signedexchange/internal/signingalgorithm package to go/internal/signingalgorithm
This package will be used from go/bundle too.
